Support decoding metadata v15
