+++ substrate-sdk-android/src/main/java/io/novasama/substrate_sdk_android/scale/Schema.kt
@@ -1,5 +1,6 @@
 package io.novasama.substrate_sdk_android.scale
 
+import android.util.Log
 import io.emeraldpay.polkaj.scale.ScaleCodecReader
 import io.emeraldpay.polkaj.scale.ScaleCodecWriter
 import io.emeraldpay.polkaj.scale.ScaleReader
@@ -56,8 +57,15 @@ abstract class Schema<S : Schema<S>> :
         val struct = EncodableStruct(this as S)
 
         for (field in fields) {
-            val value = field.dataType.read(reader)
-            struct[field as Field<Any?>] = value
+            try {
+                val value = field.dataType.read(reader)
+                struct[field as Field<Any?>] = value
+            } catch (e: Exception) {
+                println("Error reading ${this::class.simpleName}")
+
+                throw e
+            }
+
         }
 
         return struct
